Change from Astroflare to Flarekit
Update the database package structure as I was not happy with the previous implementation. A singleton instance should be created for database init along with context of cache, queue if needs to be used for caching purpose. Thus to avoid complexity of the future we have updated the code for @flarekit/database

Change cloudflare worker to Hapi worker handling the queue, schedule and api request. Change the gallery function. However this is not yet fully functional, rewamp is needed for Astro.